Implemented a basic modmail system
The user can run the command modmail in the bot's DMs they will then
be given a list of servers they (and the bot) are in. After choosing the
server based on the numbers they will be requested to send messages
which will then be relayed into the modmail channel which can be
configured using the modmail command too.