NEW: MODMAIL WITH SERVERS!!!
Extended the current modmail setup to a multiserver system with replies
which cannot be abused--checks for a valid commandCache and cross
verifies channels and users. Creates a new channel for each new modmail
and synchronises ending of sessions.
Server linkage depends on a server secret which is randomly generated
for both servers so the action they've wished to take is verified to the
greatest extent.